+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - values which enable the loading are: `"", 0, f, false, nil, none`
     - you can also specify an entirely custom path to a configuration file in this variable
 
+### Removed
+- the `INI`-style configuration got fully removed (as deprecated in [3.0.0] - 2021-04-10)
+
+
 ## [3.4.0] - 2021-12-01
 
 Pypi: https://pypi.org/project/cobib/3.4.0/